requirements.txt
phue
caldav
python-dotenv
pytz